% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_clusters.R
\name{choose_clusters}
\alias{choose_clusters}
\title{Filter clusters heuristic}
\usage{
choose_clusters(
  x,
  binomial_cutoff = 0.05,
  dimensions_cutoff = 1,
  pi_cutoff = 0.02,
  re_assign = FALSE
)
}
\arguments{
\item{x}{An object of class 'vb_bmm'.}

\item{binomial_cutoff}{Minimum Binomial success probability.}

\item{dimensions_cutoff}{Minimum number of dimensions where we want to detect a Binomial component.}

\item{pi_cutoff}{Minimum size of the mixture component.}

\item{re_assign}{If \code{TRUE}, point assigned to a cluster that is filtered our, are re-assigned
from the density function (without updating the parameters). Otherwise, clustering assignments for
those points are returned as \code{NA}.}
}
\value{
An object of class 'vb_bmm'.
}
\description{
This removes the clusters from a fit object. There are 2
conditions that can be used to filter out clusters: 1) the cluster size
(mixing proportion), which one can require to be above a certain cutoff,
and 2) the Binomial peak, whichi one can ask to be above a certain value
in at least a certain number of dimensions. Output clusters will
be renamed by size (C1 will be larger etc.), and the latent variables and
hard clustering assignments will be updated accordingly.
}
\examples{
data(fit_mvbmm_example)
choose_clusters(fit_mvbmm_example)
}
