% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_2D_binom}
\alias{example_2D_binom}
\title{Colorectal Cancer Data by Cross et al.}
\format{Data from a 2D Binomial mixture}
\usage{
data(example_2D_binom)
}
\description{
Driver mutations detected by Cross et al. in a cohort of colorecta adenomas and carcinomas.
}
\examples{
data(example_2D_binom)
head(example_2D_binom)

}
\keyword{#'}
\keyword{datasets}
