% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_mvbmm_example_4D}
\alias{fit_mvbmm_example_4D}
\title{Fit of the attached 4-dimensional Binomial mixture}
\format{Fit of a multivariate Binomial mixture with 4 dimensions}
\usage{
data(fit_mvbmm_example_4D)
}
\description{
Fir for a 4-dimensional Binomial mixture
}
\examples{
data(fit_mvbmm_example_4D)
print(fit_mvbmm_example_4D)
}
\keyword{datasets}
