% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvbmm_example}
\alias{mvbmm_example}
\title{Example data}
\format{
Data from a multivariate Binomial mixture
}
\usage{
data(mvbmm_example)
}
\description{
Data from a multivariate dimensional Binomial mixture
with unknown ground truth.
}
\examples{
data(mvbmm_example)
head(mvbmm_example)
}
\keyword{datasets}
