% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvbmm_example_4D}
\alias{mvbmm_example_4D}
\title{Example data (4-dimensional Binomial mixture)}
\format{Data from a multivariate Binomial mixture with 4 dimensions}
\usage{
data(mvbmm_example_4D)
}
\description{
Data from a 4-dimensional Binomial mixture (with unknown ground truth)
}
\examples{
data(mvbmm_example_4D)
head(mvbmm_example_4D)
}
\keyword{datasets}
