% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_plot.R
\name{plot.vb_bmm}
\alias{plot.vb_bmm}
\title{Plot the clustering assignments.}
\usage{
\method{plot}{vb_bmm}(x, ...)
}
\arguments{
\item{x}{An object of class vb_bmm.}

\item{...}{}
}
\value{
A list of plots, that can be assembled in a unique figure
with other packages like \code{ggpubr}.
}
\description{
This functions runs for every pair of dimensions
in the data the function \code{plot_2D}.
}
\examples{
data(fit_mvbmm_example)
plot(fit_mvbmm_example)
}
