% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1D.R
\name{plot_1D}
\alias{plot_1D}
\title{Plot the clustering assignments for marginal data.}
\usage{
plot_1D(x, colors = NA)
}
\arguments{
\item{x}{An object of class vb_bmm.}

\item{colors}{Optional vector of colors, default (\code{NA}) are \code{ggplot} colors.}
}
\value{
A ggplot object
}
\description{
Plot a histogram of successes over trials, coloured by clustering assignments and
with a facet for each sample.
}
\examples{
data(fit_mvbmm_example)

plot_1D(fit_mvbmm_example)
}
