% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_ELBO}
\alias{plot_ELBO}
\title{Plot the ELBO}
\usage{
plot_ELBO(x, cex = 1)
}
\arguments{
\item{x}{An object of class 'vb_bmm'.}

\item{cex}{Cex of the plot.}
}
\value{
A plot of the ELBO.
}
\description{
Plot the ELBO
}
\examples{
TODO
}
