% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_latent_variables}
\alias{plot_latent_variables}
\title{Plot the latent variables of the mixture.}
\usage{
plot_latent_variables(x, cex = 1, add.VAF = TRUE)
}
\arguments{
\item{x}{An object of class 'vb_bmm'.}

\item{cex}{Cex of the plot.}

\item{add.VAF}{Add also the VAF inside \code{x} to the plot.}
}
\value{
A plot of the latent variables of the mixture.
}
\description{
Plot the latent variables of the mixture.
}
\examples{
data(fit_mvbmm_example)
plot_latent_variables(fit_mvbmm_example)
}
