% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_variational.R
\name{vb_bmm_MV}
\alias{vb_bmm_MV}
\title{=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
log sum exp trick for having numeric stability   #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=-=-= #
Auxiliary variables   #
=-=-=-=-=-=-=-=-=-=-= #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
Bayesian priors from the hyperparameters   #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
Posterior approximations, different ways to initialize the q   #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
Normalization constants for the priors, required for the ELBO  #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
Other extra variables that we report in the fit  #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
Minimization of the KL divergence between the posterior and its approximution  #
=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=- #
Variational E-Step #
=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=- #
Variational M-Step #
=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=- #
Variational ELBO   #
=-=-=-=-=-=-=-=-=- #}
\usage{
vb_bmm_MV(x_NV, x_DP, K, alpha_0, a_0, b_0, max_iter, epsilon_conv,
  save_trace, q_init)
}
\description{
=-=-=-=-=-=-=-=-=- #
Checks convergency #
=-=-=-=-=-=-=-=-=- #
=-=-=-=-=-=-=-=-=-=-=-=-=- #
Hard clustering assignment #
=-=-=-=-=-=-=-=-=-=-=-=-=- #
}
