% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D.R
\name{plot_2D}
\alias{plot_2D}
\title{2D clusters plot.}
\usage{
plot_2D(x, d1, d2, cex = 1, alpha = 0.3, downsample = Inf,
  colors = NA)
}
\arguments{
\item{x}{An object of class 'vb_bmm'.}

\item{d1}{The name of the dimension to plot (x-axis).}

\item{d2}{The name of the dimension to plot (y-axis).}

\item{cex}{Cex of the points and the overall plot.}

\item{alpha}{Alpha of the points.}

\item{downsample}{Maximun number of points to plot. Default is \code{Inf}, no downsampling.}

\item{colors}{Optional vector of colors, default (\code{NA}) are \code{ggplot} colors.}
}
\value{

}
\description{
Plot the fit for two dimensions, one against the other. Colour each point
(defined as the ratio of sucesses over trail) according to the clustering assignments.
}
\examples{
TODO
}
