% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_fits}
\alias{plot_fits}
\title{Plot 2D and 1D fits altogether.}
\usage{
plot_fits(x, ...)
}
\arguments{
\item{x}{An object of class vb_bmm.}

\item{...}{Extra parameters for \code{plot_2D} and \code{plot_1D} functions.}
}
\value{
A figure assembled with \code{ggpubr}.
}
\description{
Performs a joint plot in which both the clustering scatter and the
marginal plots are repreoduced. This function uses the \code{plot_2D} and
\code{plot_1D} functions to create all the required plots, and the ellipsis
is used to forward parameters to both functions.
}
\examples{
data(fit_mvbmm_example)
plot_fits(fit_mvbmm_example)

\donotrun{

require(dplyr)

colors_clusters = fit_mvbmm_example$labels \%>\% unique \%>\% pull()
colors_clusters_samples = ggsci::pal_lancet()(colors_clusters \%>\% length())

names(colors_clusters_samples) = colors_clusters

plot_fits(fit_mvbmm_example, colors = colors_clusters_samples)
}
}
