% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_clusters.R
\name{rename_clusters}
\alias{rename_clusters}
\title{Rename VIBER clusters.}
\usage{
rename_clusters(x, new_labels)
}
\arguments{
\item{x}{A VIBER fit object.}

\item{new_labels}{A named vector representing a new names map. For instance
\code{c(`C1`=`Clonal`, `C2` = `Subclonal`)} renames two clusters "C1" and "C2"
as "Clonal" and "Clonal" respectively.}
}
\value{
A VIBER fit object with renamed clusters.
}
\description{
From an object fit, clusters (all or some) can be renamed.
This changes the names inside the VIBER object, affecting plots and all
other outputs. This function can be used after clusters have been given
some interpretation, e.g., cluster "C1" is the "Clonal" cluster for instance.
}
\examples{
data(mvbmm_example)
x = variational_fit(mvbmm_example$successes, mvbmm_example$trials)
print(x)

x_renamed = rename_clusters(x, c(`C1`=`Clonal`, `C2` = `Subclonal`))
print(x_renamed)
}
