% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{summary.vb_bmm}
\alias{summary.vb_bmm}
\title{Print to screen the object}
\usage{
\method{summary}{vb_bmm}(object, ...)
}
\arguments{
\item{object}{An object of class vb_bmm.}

\item{...}{Default extra paramaters.}
}
\value{
Nothing, just print to screen.
}
\description{
Print to screen the object
}
\examples{
data(fit_mvbmm_example)
print(fit_mvbmm_example)
}
