% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_variational.R
\name{variational_fit}
\alias{variational_fit}
\title{Variational fit for Binomial mixtures}
\usage{
variational_fit(x, y, K = 10, alpha_0 = 1e-06, a_0 = 1, b_0 = 1,
  max_iter = 5000, epsilon_conv = 1e-10, samples = 10,
  q_init = "prior", trace = FALSE)
}
\arguments{
\item{x}{A matrix where each column is a dimension of the multivariate Binomial,
and each row is an input point. Values of this matrix represent the number of
successes of independent Bernoulli trials. This matrix and \code{y} should have
the same dimension.}

\item{y}{A matrix where each column is a dimension of the multivariate Binomial,
and each row is an input point. Values of this matrix represent the number of
attempts of independent Bernoulli trials. This matrix and \code{x} should have
the same dimension.}

\item{K}{The maximum number of clusters returned, it should be lower than the
number of rows of \code{x} and \code{y}. Default is \code{K = 10}; lower values
speed up convergence.}

\item{alpha_0}{The concentration parameter of the Dirichlet mixture. The default
is a stringent fit with \code{alpha = 1e-6}.}

\item{a_0}{Prior Beta hyperparameter. If this values is a scalar than all the
mixture components have the same prior. The default is scalar  \code{a_0 = 1}.}

\item{b_0}{Prior Beta hyperparameter. If this values is a scalar than all the
mixture components have the same prior. The default is scalar  \code{b_0 = 1}.}

\item{max_iter}{Maximum number of fit iterations. The fit is interrupted when
this number of iterations is performed. Default \code{max_iter = 5000}}

\item{epsilon_conv}{Epsilon to measure convergence (ELBO absolute difference).}

\item{samples}{Number of fits computed by the algorithm. Only the best fit is
returned. This value must be greater or equal than 1.}

\item{q_init}{Initialization of the q-distribution to compute the approximation
of the posterior distributions. This can be set in three different waysL
equal to the prior (\code{q_init = 'prior'}), via kmeans clustering
(\code{q_init = 'kmeans'}) and capturing points which are private to each
dimension (\code{q_init = 'private'}). The default is equal to the prior.}

\item{trace}{If true the trace computed during the fit is returned (this allows
to check fits a posterirori, make animations etc.). Default is \code{FALSE}; this
feature can slow down quite substantially the fit.}
}
\value{
An object of class \code{vb_bmm} which contains S3 methods to extract
the fit, plots the results, compute summary statistics etc.
}
\description{
Variational fit for a semi-parametric Dirichelt mixture of Binomial
distributions. The fit convergency can be monitored through the ELBO,
can be run either sequentially (single core) or in parallel. You need
to provide an upper bound on the number of clusters that you want to
obtain, through parameters \code{K}. You can explicit the Dirichlet
prior for the concentration of the mixture (\code{alpha_0}), as well as
the hyperparmeters of the Beta priors for each mixture component.
}
\examples{
data(mvbmm_example)
f = variational_fit(mvbmm_example$successes, mvbmm_example$trials)
print(f)
}
